[Refactor] Split FormatAvro module
