+++ be/CMakeLists.txt
@@ -553,6 +553,7 @@ set(STARROCKS_LINK_LIBS
     Script
     orc
     ${WL_END_GROUP}
+    FormatAvro
     FormatCsv
     FormatJson
     FormatCore
