+++ be/module_boundary_manifest.json
@@ -311,7 +311,7 @@
     {
       "id": "formatcore",
       "doc_label": "FormatCore",
-      "summary": "Format-oriented core primitives above ExprCore, RuntimeCore, FSCore, ChunkCore, ColumnCore, and Types.",
+      "summary": "Format-oriented core primitives above ComputeEnv, ExprCore, RuntimeCore, FSCore, ChunkCore, ColumnCore, and Types.",
       "owned_targets": ["FormatCore"],
       "owned_globs": [
         "be/src/formats/column_evaluator.*",
@@ -337,6 +337,7 @@
         "gen_cpp/"
       ],
       "allowed_target_deps": [
+        "ComputeEnv",
         "ExprCore",
         "RuntimeCore",
         "FSCore",
@@ -350,6 +351,7 @@
       ],
       "allowed_test_targets": ["format_test"],
       "allowed_test_link_deps": [
+        "FormatAvro",
         "FormatJson",
         "FormatCsv",
         "FormatCore",
