+++ be/src/formats/avro/nullable_column.cpp
@@ -17,7 +17,6 @@
 #include "column/adaptive_nullable_column.h"
 #include "column/array_column.h"
 #include "column/nullable_column.h"
-#include "formats/json/binary_column.h"
 #include "gutil/strings/substitute.h"
 #include "types/logical_type.h"
