+++ be/test/formats/CMakeLists.txt
@@ -45,7 +45,20 @@ set(FORMAT_JSON_TEST_FILES
     json/struct_column_test.cpp
 )
 
+set(FORMAT_AVRO_TEST_FILES
+    avro/binary_column_test.cpp
+    avro/nullable_column_test.cpp
+    avro/numeric_column_test.cpp
+    avro/cpp/avro_reader_test.cpp
+    avro/cpp/binary_column_reader_test.cpp
+    avro/cpp/complex_column_reader_test.cpp
+    avro/cpp/date_column_reader_test.cpp
+    avro/cpp/numeric_column_reader_test.cpp
+    avro/cpp/utils_test.cpp
+)
+
 set(FORMAT_TEST_LINK_LIBS
+    FormatAvro
     FormatJson
     FormatCsv
     FormatCore
@@ -80,10 +93,15 @@ add_library(starrocks_format_json_test_objs OBJECT ${FORMAT_JSON_TEST_FILES})
 target_link_libraries(starrocks_format_json_test_objs PRIVATE StarRocksPCH)
 set_target_properties(starrocks_format_json_test_objs PROPERTIES COMPILE_FLAGS "-fno-access-control")
 
+add_library(starrocks_format_avro_test_objs OBJECT ${FORMAT_AVRO_TEST_FILES})
+target_link_libraries(starrocks_format_avro_test_objs PRIVATE StarRocksPCH)
+set_target_properties(starrocks_format_avro_test_objs PROPERTIES COMPILE_FLAGS "-fno-access-control")
+
 add_executable(format_test
     format_test_main.cpp
     $<TARGET_OBJECTS:starrocks_format_core_test_objs>
     $<TARGET_OBJECTS:starrocks_format_csv_test_objs>
     $<TARGET_OBJECTS:starrocks_format_json_test_objs>
+    $<TARGET_OBJECTS:starrocks_format_avro_test_objs>
 )
 target_link_libraries(format_test ${FORMAT_TEST_LINK_LIBS})
